fix: self invocation으로 인한 transaction logic 분리
